+++ custom/res/Auterion/Widgets/AuterionVehicleButton.qml
@@ -24,13 +24,17 @@ Button {
 
     function getBatteryColor() {
         if(vehicle) {
-            if(vehicle.batteryStatus.percentRemaining.value > 75) {
+            if(vehicle.batterySummary == undefined)
+                console.warn("Battery Summary: Not initialized yet")
+            else
+                console.warn("Battery Summary: OK")
+            if(vehicle.batterySummary.percentRemaining.value > 75) {
                 return qgcPal.colorGreen
             }
-            if(vehicle.batteryStatus.percentRemaining.value > 50) {
+            if(vehicle.batterySummary.percentRemaining.value > 50) {
                 return qgcPal.colorOrange
             }
-            if(vehicle.batteryStatus.percentRemaining.value > 0.1) {
+            if(vehicle.batterySummary.percentRemaining.value > 0.1) {
                 return qgcPal.colorRed
             }
         }
@@ -39,7 +43,7 @@ Button {
 
     function getBatteryPercentage() {
         if(vehicle) {
-            return vehicle.batteryStatus.percentRemaining.value / 100.0
+            return vehicle.batterySummary.percentRemaining.value / 100.0
         }
         return 1
     }