+++ custom/src/QmlControls/QGroundControl/Specific/GimbalVirtualJoystick.qml
@@ -1,6 +1,6 @@
 import QtQuick 2.0
-import QtQuick.Controls 2.5
-import QtGraphicalEffects 1.12
+import QtQuick.Controls 2.4
+import QtGraphicalEffects 1.0
 
 //import QGroundControl.Controls.VirtualJoystick
 import Auterion.Controls 1.0